client/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
Jinja2==3.1.3
grpcio==1.60.1
grpcio-tools==1.60.1
python-multipart==0.0.9
